+++ src/main/java/com/jin/project01/dto/user/MeResponse.java
@@ -7,7 +7,7 @@ import lombok.Getter;
 @Builder
 public class MeResponse {
 
-    private Long userNo;
+    private Integer userNo;
     private String userId;
     private String userName;
     private String userEmail;
