+++ src/main/java/com/jin/project01/jwt/JwtAuthenticationFilter.java
@@ -26,7 +26,7 @@ public class JwtAuthenticationFilter extends OncePerRequestFilter {
             String token = bearerToken.substring(7);
 
             if (jwtTokenProvider.validateToken(token)) {
-                Long userNo = jwtTokenProvider.getUserNo(token);
+                Integer userNo = jwtTokenProvider.getUserNo(token);
                 var userDetails = userDetailsService.loadUserByUserNo(userNo);
 
                 var authentication = new UsernamePasswordAuthenticationToken(
