+++ src/main/java/com/jin/project01/security/CustomUserDetailsService.java
@@ -24,7 +24,7 @@ public class CustomUserDetailsService implements UserDetailsService {
     }
 
     // JWT 인증 시 토큰 subject에 들어있는 userNo를 가지고 옴
-    public UserDetails loadUserByUserNo(Long userNo) throws UsernameNotFoundException {
+    public UserDetails loadUserByUserNo(Integer userNo) throws UsernameNotFoundException {
         User user = userRepository.findById(userNo)
                 .orElseThrow(() -> new UsernameNotFoundException("사용자를 찾을 수 없습니다.") );
